+++ flutter_searchbox_ui/lib/src/components/selected_filters.dart
@@ -268,7 +268,7 @@ class _SelectedFiltersState extends State<SelectedFilters> {
         // initialization block ends
         componentInstance?.subscribeToStateChanges((changes) {
           void applyChanges() {
-            final currentValue = changes['value']?.next;
+            final currentValue = changes.Value?.next;
             if (!isNullEmptyOrFalse(currentValue) &&
                 (widget.hideDefaultValues == true
                     ? !isEqual(currentValue, widget.defaultValues![id])
@@ -286,7 +286,7 @@ class _SelectedFiltersState extends State<SelectedFilters> {
           } else {
             applyChanges();
           }
-        }, ["value"]);
+        }, [KeysToSubscribe.Value]);
       }
     } catch (e) {
       print('error $e');