+++ searchbase/lib/src/observer.dart
@@ -1,6 +1,8 @@
 /// Represents an observer function that listens for the state changes.
+import 'types.dart';
+
 class Observer {
   Function callback;
-  List<String>? properties;
+  List<KeysToSubscribe>? properties;
   Observer(this.callback, this.properties) {}
 }