+++ flutter_searchbox/CHANGELOG.md
@@ -1,3 +1,23 @@
+## [3.0.0] - 21-04-2022
+
+- Add `StateProvider()` widget.
+- Breaking Change (as per **searchbase: 3.0.0**)
+    - SearchWidgetConnector/ SearchController accepts subscribeTo as `List<KeysToSubscribe>`;
+    - `subscribeToStateChanges()` now accepts array of enums (`KeysToSubscribe`) for properties to subscribe, instead of hard-coded strings.
+        Also, the subscription callback function receives a `ChangesController` class argument instead of `Changes` type.
+        - **before**
+        ```dart
+            widgetInstance.subscribeToStateChanges((Changes change){
+                print('${changes['results']!.next}');
+            }, ['results']);
+        ```
+        - **after**
+        ```dart
+            widgetInstance.subscribeToStateChanges((ChangesController change){
+                print('${changes.Results!.next}');
+            }, [KeysToSubscribe.Results]);
+        ```        
+    > Refer to docs to see the `KeysToSubscribe` enum defined under *types.dart*
 ## [2.2.4-nullsafety] - 15-01-2022
 
 - Update searchbase